feat(request): Streamlining
